Phase 0: a destruction harness that runs without a GPU, and CI that runs it (#29)
* Add the destruction and particle simulation rewrite plan

* Phase 0: a destruction harness that runs without a GPU, and CI that runs it

Every later phase of DESTRUCTION_PLAN.md claims a perf win or behavioural
identity. Without something to measure against, those claims are vibes. This
builds the instrument.

The gauntlet is a CPU harness rather than a scripted run of the game, which is
a deliberate departure from the plan's work item 3. Scripting the engine cannot
be made deterministic - a real level is full of entities with their own
randomness, chunk streaming runs on job threads, the camera slides the window,
and the frame rate decides how many fixed ticks a second holds - and a state
hash without determinism is worthless. What makes the alternative possible is
rule 1: everything destruction touches is ordinary memory except the mapped
voxel buffer, and the CPU only ever writes that, so handing the write path a
plain vector<uint32_t> reproduces it at full fidelity with no GPU, no window
and no World. That is VoxelWorldHarness, and it is why the gauntlet and the
unit tests run in CI.

Its reference writer is the point, not scaffolding: it is the dumbest possible
implementation of rule 3 - touch every representation, in the open, no
batching. Phase 1's VoxelEditBatch gets checked against it by running the same
edit script through both, which says the new path agrees with an independent
implementation rather than with its own previous self.

CI now builds the engine, the game, the bring-up target and the tests on both
configurations and runs ctest. VOXAGINE_BUILD_ENGINE had been off there for the
whole port, so the only C++ it compiled was VulkanBringup.cpp. Turning it on
needed one fix: add_dependencies on the shader target is a generate-time error
when dxc is absent, and the runner has no Vulkan SDK.

VOXAGINE_SYNC_AUDIT=<seconds> is the in-game half - it compares the CPU voxel
against the mapped word, which is the one link ValidateBrickGrid cannot see. It
stages the mapping with one bulk copy first; reading it in place per voxel is
minutes per pass and the first version simply never returned. Even staged the
read is 19 s on a 768^3 window, so it wants an interval of a minute. Its first
run reports zero brick, bitmap and occupied-only-on-the-CPU disagreements.

Particle velocities come from a seeded xorshift64* rather than glm::linearRand,
which draws on a process-global engine every other caller perturbs - a replay
diverged on the first unrelated call anywhere. That is the one behavioural
change here and it is invisible in play.

Five integer class constants became static constexpr. static const with an
in-class initializer is odr-used by anything taking it by reference, which a
gtest EXPECT_EQ does; this is the third Release-only link difference of that
exact shape in this tree.

Closes ledger M2: the deleted IntegrityJob is gone from CLAUDE.md, SOURCE_MAP,
ARCHITECTURE and the findings index, where VX-PHY-002/003 are now
overtaken-by-events and VX-PHY-001 partially-overtaken under a new resolution
field.

* Use the unsuffixed std:: math names, which exist

std::cosf, std::sinf and std::fmodf are not standard: the C functions are only
guaranteed in the global namespace, and whether <cmath> also drags them into
std is a libstdc++ version detail. This machine's does and the CI runner's does
not, so eight call sites compiled here and failed there.

The float overload of std::cos is the same function, so nothing changes but the
spelling. Found by turning VOXAGINE_BUILD_ENGINE on in CI, which had never
compiled either of these files.

---------

+++ Game/Source/Weapons/Bullet.cpp
@@ -264,11 +264,11 @@ void Bullet::FixedTick(const GameTimer& gameTimer)
 
 						if (m_fPathAnimationTimer >= m_fPathAnimationTime)
 						{
-							m_fPathAnimationTimer = std::fmodf(m_fPathAnimationTimer + fDeltaTime, m_fPathAnimationTime);
+							m_fPathAnimationTimer = std::fmod(m_fPathAnimationTimer + fDeltaTime, m_fPathAnimationTime);
 							m_uiPathAnimationLoopCount++;
 						}
 
-						m_AnimationOffset = m_Right * std::sinf(-PI + m_fPathAnimationTimer / m_fPathAnimationTime * PI * 2.f);
+						m_AnimationOffset = m_Right * std::sin(-PI + m_fPathAnimationTimer / m_fPathAnimationTime * PI * 2.f);
 
 						break;
 					}
